64 bits port: updated data type to avoid warnings.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@46735 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/console/argcargv.c
@@ -30,7 +30,7 @@ rt_private void shword(char *cmd, int *argc, char ***argvp);
 
 rt_public void get_argcargv (int *argc, char ***argv)
 {
-	int tl;
+	size_t tl;
 	temp = strdup (GetCommandLine());
 
 		/* Only for Application that are launched from EiffelBench:
